refactor: split MCP allowance tools
