+++ src/codex_usage_tracker/cli/mcp_runtime.py
@@ -1,0 +1,5 @@
+"""Shared MCP registration runtime."""
+
+from mcp.server.fastmcp import FastMCP
+
+mcp = FastMCP("codex-usage-tracker")
